Rename and Move CommandRanEvent to be more obvious as to what it does, and create package-info.java
