+++ api/src/main/java/com/velocitypowered/api/event/command/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides events for handling commands.
+ */
+package com.velocitypowered.api.event.command;
